Refactor codebase by moving common-code to open-api-service module
