+++ openapi-cli/src/main/java/io/ballerina/openapi/generators/openapi/OpenApiConverter.java
@@ -20,8 +20,8 @@ package io.ballerina.openapi.generators.openapi;
 
 import io.ballerina.compiler.api.SemanticModel;
 import io.ballerina.compiler.syntax.tree.SyntaxTree;
-import io.ballerina.openapi.cmd.utils.CodegenUtils;
 import io.ballerina.openapi.converter.OpenApiConverterException;
+import io.ballerina.openapi.converter.utils.CodegenUtils;
 import io.ballerina.openapi.converter.utils.ServiceToOpenAPIConverterUtils;
 import io.ballerina.projects.Document;
 import io.ballerina.projects.DocumentId;
